refactor: Request/Response 클래스 domain 패키지에서 분리
