+++ src/main/java/roomescape/dto/reservation/ReservationRequest.java
@@ -1,4 +1,4 @@
-package roomescape.domain.reservation;
+package roomescape.dto.reservation;
 
 import roomescape.exception.InvalidInputException;
