+++ src/main/java/roomescape/dto/theme/ThemeRequest.java
@@ -1,4 +1,4 @@
-package roomescape.domain.theme;
+package roomescape.dto.theme;
 
 public record ThemeRequest(String name, String description, String url) {
