+++ src/main/java/roomescape/repository/ThemeUpdatingDao.java
@@ -4,7 +4,7 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
-import roomescape.domain.theme.ThemeRequest;
+import roomescape.dto.theme.ThemeRequest;
 
 import java.sql.PreparedStatement;
